Paranoid Backup #1
A bunch of unimplemented features, but I'm too cautious to do this much
without backing up.

If you run this as is, you'll see a noticeable difference since the last
commit to master. Check next non-backup commit to see the changes.

+++ Last Prototype/Last Prototype/Sprite.h
@@ -7,7 +7,7 @@ A sprite is an EntityContainer that loops through an animation based on a cycle.
 class Sprite : public EntityContainer
 {
 public:
-	Sprite(int x, int y) : EntityContainer(x, y), cycle(0) {}
+	Sprite(float x, float y) : EntityContainer(x, y), cycle(0) {}
 	~Sprite(void) {}
 	void render(void) { set_skin(); EntityContainer::render(); }
